requirements.txt
numpy
matplotlib
argparse
cython
pysoundfile
pyworld
